试卷翻新多智能体/requirements.txt
# ===========================================
# 医学试卷翻新多智能体系统 - 依赖列表
# ===========================================

# 数据模型与验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 文件解析
python-docx>=1.1.0
pypdf>=4.0.0

# 向量数据库与嵌入
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM 集成
openai>=1.0.0
google-generativeai>=0.8.0

# 异步支持
aiofiles>=23.0.0

# 环境变量管理
python-dotenv>=1.0.0

# 进度条与日志
rich>=13.0.0
tqdm>=4.65.0

# 重试机制
tenacity>=8.2.0

# 文本处理
tiktoken>=0.5.0

# Web UI
gradio>=4.0.0
